src/main/java/com/codechallenge/commitviewer/infrastructure/rest/json/Author.java
package com.codechallenge.commitviewer.infrastructure.rest.json;

import java.time.LocalDateTime;

public class Author {

    private String name;
    private String email;
    private LocalDateTime date;

    public String getName() {
        return name;
    }

    public void setName(String name) {
        this.name = name;
    }

    public String getEmail() {
        return email;
    }

    public void setEmail(String email) {
        this.email = email;
    }

    public LocalDateTime getDate() {
        return date;
    }

    public void setDate(LocalDateTime date) {
        this.date = date;
    }

}
